# ranking: Speed up dominating SQL query